Initial separation of zoom functionality for reuse.
