+++ src/main/java/org/tros/torgo/swing/ZoomableComponent.java
@@ -1,0 +1,175 @@
+/*
+ * Copyright 2015-2017 Matthew Aguirre
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package org.tros.torgo.swing;
+
+import java.awt.event.ActionEvent;
+import java.awt.event.InputEvent;
+import java.awt.event.KeyEvent;
+import java.awt.event.KeyListener;
+import java.awt.event.MouseEvent;
+import java.awt.event.MouseListener;
+import java.awt.event.MouseWheelEvent;
+import java.lang.reflect.InvocationTargetException;
+import java.lang.reflect.Method;
+import java.util.concurrent.atomic.AtomicBoolean;
+import javax.swing.JComponent;
+import javax.swing.JMenuItem;
+import javax.swing.JPopupMenu;
+import javax.swing.KeyStroke;
+import org.tros.utils.logging.Logger;
+import org.tros.utils.logging.Logging;
+
+/**
+ *
+ * @author matta
+ */
+public abstract class ZoomableComponent {
+
+    private static final Logger LOGGER = Logging.getLogFactory().getLogger(ZoomableComponent.class);
+
+    public ZoomableComponent(JComponent component) {
+        Runnable increase = () -> {
+            LOGGER.info("increase");
+            zoomIn();
+        };
+
+        Runnable decrease = () -> {
+            LOGGER.info("decrease");
+            zoomOut();
+        };
+
+        Runnable reset = () -> {
+            LOGGER.info("reset");
+            zoomReset();
+        };
+
+        final AtomicBoolean ctrlDown = new AtomicBoolean(false);
+        component.addKeyListener(new KeyListener() {
+            @Override
+            public void keyTyped(KeyEvent ke) {
+            }
+
+            @Override
+            public void keyPressed(KeyEvent ke) {
+                if ((ke.getModifiers() & KeyEvent.CTRL_MASK) != 0) {
+                    ctrlDown.set(true);
+                }
+                if ((ke.getKeyCode() == KeyEvent.VK_EQUALS)
+                        && (ke.getModifiers() == (KeyEvent.CTRL_MASK | KeyEvent.SHIFT_MASK))
+                        || (ke.getKeyCode() == KeyEvent.VK_ADD)
+                        && (ke.getModifiers() == (KeyEvent.CTRL_MASK))) {
+                    increase.run();
+                }
+                if ((ke.getKeyCode() == KeyEvent.VK_MINUS || ke.getKeyCode() == KeyEvent.VK_SUBTRACT)
+                        && ((ke.getModifiers() == KeyEvent.CTRL_MASK))) {
+                    decrease.run();
+                }
+                if ((ke.getKeyCode() == KeyEvent.VK_0 || ke.getKeyCode() == KeyEvent.VK_NUMPAD0)
+                        && ((ke.getModifiers() == KeyEvent.CTRL_MASK))) {
+                    reset.run();
+                }
+            }
+
+            @Override
+            public void keyReleased(KeyEvent ke) {
+                if (ke.getModifiers() == 0) {
+                    ctrlDown.set(false);
+                }
+            }
+        });
+
+        component.addMouseWheelListener((MouseWheelEvent mwe) -> {
+            if (ctrlDown.get()) {
+                if (mwe.getPreciseWheelRotation() < 0) {
+                    increase.run();
+                } else {
+                    decrease.run();
+                }
+            }
+        });
+        JMenuItem jmi1 = new JMenuItem("Zoom In");
+        jmi1.addActionListener((ActionEvent ae) -> {
+            increase.run();
+        });
+        JMenuItem jmi2 = new JMenuItem("Zoom Out");
+        jmi2.addActionListener((ActionEvent ae) -> {
+            decrease.run();
+        });
+        JMenuItem jmi3 = new JMenuItem("Zoom Reset");
+        jmi3.addActionListener((ActionEvent ae) -> {
+            reset.run();
+        });
+        jmi1.setAccelerator(KeyStroke.getKeyStroke(KeyEvent.VK_PLUS, InputEvent.CTRL_MASK));
+        jmi2.setAccelerator(KeyStroke.getKeyStroke(KeyEvent.VK_MINUS, InputEvent.CTRL_MASK));
+        jmi3.setAccelerator(KeyStroke.getKeyStroke(KeyEvent.VK_0, InputEvent.CTRL_MASK));
+        final AtomicBoolean added = new AtomicBoolean(false);
+        final Runnable r = () -> {
+
+            if (!added.get()) {
+                added.set(true);
+                JPopupMenu popupMenu;
+                try {
+                    Method method = component.getClass().getMethod("getPopupMenu");
+                    popupMenu = (JPopupMenu)method.invoke(component);
+                } catch (NoSuchMethodException | SecurityException | IllegalAccessException | IllegalArgumentException | InvocationTargetException ex) {
+                    popupMenu = component.getComponentPopupMenu();
+                }
+                popupMenu = popupMenu == null ? new JPopupMenu() : popupMenu;
+                popupMenu.add(jmi1);
+                popupMenu.add(jmi2);
+                popupMenu.add(jmi3);
+                component.setComponentPopupMenu(popupMenu);
+            }
+        };
+        component.addMouseListener(new MouseListener() {
+            @Override
+            public void mouseClicked(MouseEvent me) {
+                r.run();
+                component.removeMouseListener(this);
+            }
+
+            @Override
+            public void mousePressed(MouseEvent me) {
+                r.run();
+                component.removeMouseListener(this);
+            }
+
+            @Override
+            public void mouseReleased(MouseEvent me) {
+                r.run();
+                component.removeMouseListener(this);
+            }
+
+            @Override
+            public void mouseEntered(MouseEvent me) {
+                r.run();
+                component.removeMouseListener(this);
+            }
+
+            @Override
+            public void mouseExited(MouseEvent me) {
+                r.run();
+                component.removeMouseListener(this);
+            }
+        });
+    }
+
+    protected abstract void zoomIn();
+
+    protected abstract void zoomOut();
+
+    protected abstract void zoomReset();
+}
